chore(refresh): refactor
