+++ src/main/kotlin/com/github/gplassard/managedexcludes/actions/AddManagedExcludesAction.kt
@@ -1,40 +1,15 @@
 package com.github.gplassard.managedexcludes.actions
 
-import com.github.gplassard.managedexcludes.services.ConfigService
 import com.github.gplassard.managedexcludes.services.ExcludeService
-import com.intellij.openapi.actionSystem.AnAction
-import com.intellij.openapi.actionSystem.AnActionEvent
-import com.intellij.openapi.actionSystem.PlatformCoreDataKeys
-import com.intellij.openapi.components.service
 import com.intellij.openapi.diagnostic.thisLogger
-import kotlinx.coroutines.*
+import com.intellij.openapi.module.Module
+import com.intellij.openapi.vfs.VirtualFile
 
 
-class AddManagedExcludesAction : AnAction() {
-    override fun actionPerformed(event: AnActionEvent) {
-        thisLogger().info("START")
-        val module = event.getData(PlatformCoreDataKeys.MODULE)
-        if (module == null) {
-            thisLogger().warn("No module, aborting")
-            return
-        }
-        val excludeService = event.project?.service<ExcludeService>()
-        if (excludeService == null) {
-            thisLogger().warn("No exclude service, aborting")
-            return
-        }
-        val configService = event.project?.service<ConfigService>()
-        if (configService == null) {
-            thisLogger().warn("No config service, aborting")
-            return
-        }
-        // FIXME
-        GlobalScope.launch {
-            val excludes = configService.loadExcludeConfig(module)
-            thisLogger().warn("Going to exclude ${excludes.size} files")
-            excludeService.excludePaths(module, excludes)
-            thisLogger().info("END")
-        }
-
+class AddManagedExcludesAction : BaseManagedExcludesAction() {
+    override suspend fun perform(excludeService: ExcludeService, module: Module, excludes: Set<VirtualFile>) {
+        thisLogger().warn("Going to exclude ${excludes.size} files")
+        excludeService.excludePaths(module, excludes)
+        thisLogger().info("END")
     }
 }
